+++ weblate/utils/tests/test_search.py
@@ -274,7 +274,9 @@ class QueryParserTest(TestCase, SearchMixin):
         self.assert_query("has:dismissed-check", Q(check__dismissed=True))
         self.assert_query("has:translation", Q(state__gte=STATE_TRANSLATED))
         self.assert_query("has:variant", Q(variant__isnull=False))
-        self.assert_query("has:label", Q(source_unit__labels__isnull=False))
+        self.assert_query(
+            "has:label", Q(source_unit__labels__isnull=False) | Q(labels__isnull=False)
+        )
         self.assert_query("has:context", ~Q(context=""))
         self.assert_query(
             "has:screenshot",
@@ -323,7 +325,8 @@ class QueryParserTest(TestCase, SearchMixin):
     def test_labels(self):
         self.assert_query(
             "label:'test label'",
-            Q(source_unit__labels__name__iexact="test label"),
+            Q(source_unit__labels__name__iexact="test label")
+            | Q(labels__name__iexact="test label"),
         )
 
     def test_priority(self):